pyproject.toml
[tool.poetry]
name = "tau-vehicle"
version = "0.1.0"
description = "A package to train and infer the solution on \"TAU Vehicle Type Recognition Competition\" competition on Kaggle."
authors = ["Kirill Kondrashov <kirpall@ya.ru>"]
readme = "README.md"
packages = [
    { include = "tau_vehicle" }
]

[tool.poetry.dependencies]
python = "^3.10"
dvc = "^3.59.2"
mlflow = "^2.22.0"
dvc-s3 = "^3.2.0"
hydra-core = "^1.3.2"
fire = "^0.7.0"
pyopenssl = "^25.1.0"
pytorch-lightning = "^2.5.1.post0"
transformers = "^4.52.4"
torchvision = "^0.22.1"
python-dotenv = "^1.1.0"
timm = "^1.0.15"
onnx = "^1.18.0"
tensorrt = "^10.11.0.33"
onnxruntime = "^1.22.0"
opencv-python = "^4.11.0.86"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
